feature-structure-base-data
I made a Scriptable Object based on the docs we have about the game to store the data that the structures should have.

This means that we can easily create a new structure object inside of Unity and set its types, their sprite, resource effects, cost to make, and more.